leaderboard/requirements.txt
Django==3.2.6
django-cors-headers==3.7.0
djangorestframework==3.12.4
psycopg2-binary==2.9.1
psycopg==3.2.4
python-dotenv==1.0.1
celery==5.3.6
redis==5.0.1
gunicorn==21.2.0
django-celery-beat==2.7.0